Field "topicId" in DirectMessagesTopic updated to type Long

+++ library/src/main/java/com/pengrad/telegrambot/model/DirectMessagesTopic.kt
@@ -2,7 +2,7 @@ package com.pengrad.telegrambot.model
 
 @Suppress("unused")
 data class DirectMessagesTopic (
-    @get:JvmName("topicId") val topicId: Int,
+    @get:JvmName("topicId") val topicId: Long,
     @get:JvmName("user") val user: User?
 ) {
 }
